+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.8] - 2019-08-08
+### Changed
+- Service deployment - Agent selection: workflow updated; the LM calls the UM before calling the QoS Provider
+
+### Fixed
+- create agreement path was fixed
+
 ## [1.3.7] - 2019-09-17
 ### Added
 - new service added to check health of components: UM and LM